pyproject.toml
[project]
name = "timeline-extraction"
version = "0.1.0"
description = "Temporal Relation Classification System for Timeline Extraction using Large Language Models"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas",
    "beautifulsoup4",
    "matplotlib",
    "transformers",
    "networkx",
    "lxml",
    "google-generativeai",
    "tiktoken",
    "scikit-learn",
    "ruff>=0.11.2",
    "click",
    "torch",
    "openai",
    "python-dotenv",
    "tqdm",
    "groq",
    "together",
    "pydantic",
    "pyyaml",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
    "--tb=short",
    "--cov=timeline_extraction",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml"
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests"
]

[dependency-groups]
dev = [
    "black>=25.1.0",
]
